pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ReqCheck"
version = "0.1.0"
description = "ReqCheck - Software Requirements Checking"
authors = [
  {name = "Glaucia Schnoeller"},
]
requires-python = ">=3.8"

dependencies = [
  # Backend
  "Django>=4.2",
  "gunicorn",

  # Banco de dados
  "psycopg[binary]",
  "pgvector",

  # API
  "djangorestframework",
  "python-decouple",

  # LangChain / RAG
  "langchain",  # versão estável
  "langchain-community",
  "langchain-huggingface",
  "sentence-transformers",
  "unstructured[all-docs]",  # versão testada com LangChain
  "PyPDF2",

  # Outras libs opcionais que podem ajudar
  "tqdm",
  "requests",
  "pydantic"  # garante compatibilidade com LangChain
]

[tool.setuptools]
package-dir = {"" = "rag_engine"}

[tool.setuptools.packages.find]
where = ["rag_engine"]

[tool.isort]
profile = "black"
line_length = 100

[tool.flake8]
max-line-length = 100
